Refactor config package, deprecate -ddconfig
Properly distinguish the three sources of config (ini,
yaml, env) and merge them in a clear order.
Make it clear that -ddconfig is a deprecated option.
